requirements.txt
RPi.GPIO
Adafruit_GPIO
picamera
edge_impulse_linux
numpy
pyaudio
luma.oled
Pillow